requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
boto3==1.36.6
botocore==1.36.6
branca==0.8.1
CacheControl==0.14.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
fastapi==0.115.7
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
folium==0.19.4
fonttools==4.55.6
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
iso3166==2.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.8
mapbox==0.18.1
MarkupSafe==3.0.2
matplotlib==3.10.0
msgpack==1.1.0
networkx==3.4.2
numpy==2.2.2
openai==1.60.1
osmnx==2.0.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
polyline==2.0.2
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.7.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
s3transfer==0.11.2
scikit-learn==1.6.1
scipy==1.15.1
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
xyzservices==2025.1.0
